organize files, i18n, sort generated sets by num of matched perks, add exclude item
