+++ src/app/d2-loadout-builder/generated-sets/PlugTooltip.tsx
@@ -1,0 +1,35 @@
+import * as React from 'react';
+import { D2Item, DimSocket } from '../../inventory/item-types';
+
+export default function PlugTooltip({ item, socket }: { item: D2Item; socket: DimSocket }) {
+  const plug = socket.plug;
+  if (!plug) {
+    return null;
+  }
+
+  return (
+    <>
+      <h2>
+        {plug.plugItem.displayProperties.name}
+        {item.masterworkInfo &&
+          plug.plugItem.investmentStats &&
+          plug.plugItem.investmentStats[0] &&
+          item.masterworkInfo.statHash === plug.plugItem.investmentStats[0].statTypeHash &&
+          ` (${item.masterworkInfo.statName})`}
+      </h2>
+
+      {plug.plugItem.displayProperties.description ? (
+        <div>{plug.plugItem.displayProperties.description}</div>
+      ) : (
+        plug.perks.map((perk) => (
+          <div key={perk.hash}>
+            {plug.plugItem.displayProperties.name !== perk.displayProperties.name && (
+              <div>{perk.displayProperties.name}</div>
+            )}
+            <div>{perk.displayProperties.description}</div>
+          </div>
+        ))
+      )}
+    </>
+  );
+}
