+++ src/app/d2-loadout-builder/loadoutbuilder.scss
@@ -7,112 +7,6 @@
   flex-direction: column;
 }
 
-.locked-item {
-  cursor: pointer;
-  position: relative;
-  text-align: center;
-  width: calc(var(--item-size) + 15px);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  .empty-item {
-    margin: 2px;
-    background-color: #656565;
-    width: var(--item-size);
-    height: var(--item-size);
-    border: 2px solid #ddd;
-  }
-}
-
-.add-perk-options {
-  width: 300px;
-  height: auto;
-  flex-wrap: wrap;
-  background: #222;
-  position: absolute;
-  z-index: 1;
-  top: 44px;
-  left: 6px;
-}
-.add-perk-options-title {
-  margin: 5px 0;
-  display: flex;
-  .close {
-    top: 5px;
-    right: 5px;
-  }
-}
-.add-perk-options-details {
-  text-align: left;
-  padding: 5px;
-}
-.add-perk-options-content {
-  cursor: pointer;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.perk-image {
-  width: 40px;
-  height: 40px;
-  border: 1px solid #222;
-  // width: calc(var(--item-size) * 15px);
-  // height: calc(var(--item-size) * 15px);
-  &:hover {
-    border: 1px solid gray;
-  }
-}
-.perk-image.locked-perk {
-  border: 1px solid orange;
-}
-
-.add-perk-container {
-  opacity: 0.5;
-  display: table-cell;
-  vertical-align: middle;
-  width: var(--item-size);
-  height: var(--item-size);
-
-  &:hover {
-    opacity: 1;
-    transition: 0.3s;
-  }
-
-  &.has-locked {
-    opacity: 1;
-  }
-
-  .add-perk-text {
-    display: block;
-    font-size: 8px;
-  }
-}
-
-.add-perk-label {
-  font-size: 10px;
-}
-
-.on-loadout-hover {
-  .empty-item {
-    background-color: limegreen;
-  }
-}
-.close {
-  z-index: 1;
-}
-
-.clear {
-  background-color: rgba(100, 100, 100, 0.6);
-  position: absolute;
-  right: 5px;
-  border: none;
-  color: white;
-  &:hover {
-    background-color: rgba(0, 0, 0, 0.6);
-  }
-}
-
 .generated-build {
   padding-bottom: 20px;
