+++ src/app/d2-loadout-builder/locked-armor/lockeditem.scss
@@ -1,0 +1,17 @@
+.locked-item {
+  cursor: pointer;
+  position: relative;
+  text-align: center;
+  width: calc(var(--item-size) + 15px);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  .empty-item {
+    margin: 2px;
+    background-color: #656565;
+    width: var(--item-size);
+    height: var(--item-size);
+    border: 2px solid #ddd;
+  }
+}
