+++ src/app/d2-loadout-builder/locked-armor/popup/LockablePerksTab.tsx
@@ -1,6 +1,7 @@
-import * as React from 'react';
-import { LockType } from './LoadoutBuilder';
 import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
+import { t } from 'i18next';
+import * as React from 'react';
+import { LockType } from '../../types';
 import SelectableBungieImage from './SelectableBungieImage';
 
 export default function LockablePerks({
@@ -8,8 +9,7 @@ export default function LockablePerks({
   locked,
   hoveredPerk,
   onPerkHover,
-  toggleLockedPerk,
-  reset
+  toggleLockedPerk
 }: {
   perks: DestinyInventoryItemDefinition[];
   locked?: LockType;
@@ -19,7 +19,6 @@ export default function LockablePerks({
   };
   onPerkHover(hoveredPerk?: {}): void;
   toggleLockedPerk(perk: DestinyInventoryItemDefinition): void;
-  reset(): void;
 }) {
   const isLocked = locked && locked.items && locked.items.length !== 0;
 
@@ -33,13 +32,7 @@ export default function LockablePerks({
 
   return (
     <>
-      {isLocked && (
-        <button className="clear" onClick={reset}>
-          Reset
-        </button>
-      )}
-
-      <div>Select perks to lock</div>
+      <div>{t('LoadoutBuilder.LockPerksTitle')}</div>
       <div className="add-perk-options-content" onMouseLeave={resetHover}>
         {perks &&
           perks.map((perk) => (
